Moves custom UI classes to seperate file
